Moves payload API schemas into a dedicated module
